chore(localized-money-input): move to packages
