+++ packages/components/inputs/localized-money-input/src/currency-control.js
