+++ packages/components/inputs/localized-money-input/src/index.js
@@ -1,1 +1,3 @@
 export { default } from './localized-money-input';
+
+export { default as version } from './version';
