+++ packages/components/inputs/localized-money-input/src/localized-money-input.js
