+++ packages/components/inputs/localized-money-input/src/localized-money-input.spec.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { render, fireEvent } from '../../../../test/test-utils';
+import { render, fireEvent } from '../../../../../test/test-utils';
 import LocalizedMoneyInput from './localized-money-input';
 
 // We use this component to simulate the whole flow of
