+++ packages/components/inputs/localized-money-input/src/localized-money-input.visualspec.js
