+++ packages/components/inputs/localized-money-input/src/version.js
